Enhanced config management

+++ src/commands/remove.rs
@@ -2,15 +2,8 @@ use crate::config::Config;
 use std::env::current_dir;
 
 pub fn remove(name: &String) {
-    let current_dir = current_dir().unwrap();
-
-    if !current_dir.join(".dotfm").exists() {
-        eprintln!("Error: Not in a dotfm repository.");
-        std::process::exit(1);
-    }
-
-    let mut config_content = std::fs::read_to_string(current_dir.join(".dotfm")).unwrap();
-    let mut config: Config = toml::from_str(&config_content).unwrap();
+    let config_file = current_dir().unwrap().join(".dotfm");
+    let mut config = Config::load(&config_file).unwrap();
 
     if !config.files.contains_key(name) {
         eprintln!("Error: No managed file with this name.");
@@ -18,13 +11,15 @@ pub fn remove(name: &String) {
     }
 
     let original_path = config.files.get(name).unwrap().clone();
-    std::fs::remove_file(&original_path).unwrap();
-    std::fs::rename(current_dir.join(name), &original_path).unwrap();
+    if original_path.exists() {
+        std::fs::remove_file(&original_path).unwrap();
+    }
+
+    std::fs::rename(config_file.parent().unwrap().join(name), &original_path).unwrap();
 
     config.files.remove(name);
 
-    config_content = toml::to_string(&config).unwrap();
-    match std::fs::write(current_dir.join(".dotfm"), config_content) {
+    match config.save(&config_file) {
         Ok(_) => println!("Removed {} from {} repository.", name, config.name),
         Err(_) => {
             eprintln!("Error: Couldn't update .dotfm file.");
